Feature added: Create account and log in capability added
Server side code has been edited to contain the proper sql calls to search for specific emails or add in new emails and passwords.

Log in requests the password hash of the specific email and checks to see if they match. If they do, the user is logged in.

Log out feature also added. Once logged out you can either log in or create a new account.